All repos are recorded. better logging
Added skip reasons to repos. All repos are save. If skipped, a reason is
recorded.
When an error happens, when logging the error the severity level is
recorded along the repo, when it applies. This way we could revisit the
repo later, in case of severe exception. (perhaps for file not found
too).